# productCenter/api/bff/v1/bff_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.35.0
// source: bff/v1/bff.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBFFGetProductDetail = "/bff.v1.BFF/GetProductDetail"
const OperationBFFGetShopHome = "/bff.v1.BFF/GetShopHome"
const OperationBFFListProducts = "/bff.v1.BFF/ListProducts"

type BFFHTTPServer interface {
	// GetProductDetail 商品详情页 — 聚合商品 + 店铺 + 标签 + SKU + 副图
	GetProductDetail(context.Context, *GetProductDetailRequest) (*GetProductDetailResponse, error)
	// GetShopHome 店铺主页 — 聚合店铺 + 商品列表
	GetShopHome(context.Context, *GetShopHomeRequest) (*GetShopHomeResponse, error)
	// ListProducts 商品列表页 — 聚合商品 + 店铺名 + 标签
	ListProducts(context.Context, *ListProductsRequest) (*ListProductsResponse, error)
}

func RegisterBFFHTTPServer(s *http.Server, srv BFFHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/bff/products/{id}", _BFF_GetProductDetail0_HTTP_Handler(srv))
	r.GET("/api/v1/bff/products", _BFF_ListProducts1_HTTP_Handler(srv))
	r.GET("/api/v1/bff/shops/{id}", _BFF_GetShopHome0_HTTP_Handler(srv))
}

func _BFF_GetProductDetail0_HTTP_Handler(srv BFFHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetProductDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBFFGetProductDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetProductDetail(ctx, req.(*GetProductDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetProductDetailResponse)
		return ctx.Result(200, reply)
	}
}

func _BFF_ListProducts1_HTTP_Handler(srv BFFHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListProductsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBFFListProducts)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListProducts(ctx, req.(*ListProductsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListProductsResponse)
		return ctx.Result(200, reply)
	}
}

func _BFF_GetShopHome0_HTTP_Handler(srv BFFHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetShopHomeRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBFFGetShopHome)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetShopHome(ctx, req.(*GetShopHomeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetShopHomeResponse)
		return ctx.Result(200, reply)
	}
}

type BFFHTTPClient interface {
	// GetProductDetail 商品详情页 — 聚合商品 + 店铺 + 标签 + SKU + 副图
	GetProductDetail(ctx context.Context, req *GetProductDetailRequest, opts ...http.CallOption) (rsp *GetProductDetailResponse, err error)
	// GetShopHome 店铺主页 — 聚合店铺 + 商品列表
	GetShopHome(ctx context.Context, req *GetShopHomeRequest, opts ...http.CallOption) (rsp *GetShopHomeResponse, err error)
	// ListProducts 商品列表页 — 聚合商品 + 店铺名 + 标签
	ListProducts(ctx context.Context, req *ListProductsRequest, opts ...http.CallOption) (rsp *ListProductsResponse, err error)
}

type BFFHTTPClientImpl struct {
	cc *http.Client
}

func NewBFFHTTPClient(client *http.Client) BFFHTTPClient {
	return &BFFHTTPClientImpl{client}
}

// GetProductDetail 商品详情页 — 聚合商品 + 店铺 + 标签 + SKU + 副图
func (c *BFFHTTPClientImpl) GetProductDetail(ctx context.Context, in *GetProductDetailRequest, opts ...http.CallOption) (*GetProductDetailResponse, error) {
	var out GetProductDetailResponse
	pattern := "/api/v1/bff/products/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBFFGetProductDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetShopHome 店铺主页 — 聚合店铺 + 商品列表
func (c *BFFHTTPClientImpl) GetShopHome(ctx context.Context, in *GetShopHomeRequest, opts ...http.CallOption) (*GetShopHomeResponse, error) {
	var out GetShopHomeResponse
	pattern := "/api/v1/bff/shops/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBFFGetShopHome))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListProducts 商品列表页 — 聚合商品 + 店铺名 + 标签
func (c *BFFHTTPClientImpl) ListProducts(ctx context.Context, in *ListProductsRequest, opts ...http.CallOption) (*ListProductsResponse, error) {
	var out ListProductsResponse
	pattern := "/api/v1/bff/products"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBFFListProducts))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}