# productCenter/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/v1/bff/products:
        get:
            tags:
                - BFF
            description: 商品列表页 — 聚合商品 + 店铺名 + 标签
            operationId: BFF_ListProducts
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: shopId
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/bff.v1.ListProductsResponse'
    /api/v1/bff/products/{id}:
        get:
            tags:
                - BFF
            description: 商品详情页 — 聚合商品 + 店铺 + 标签 + SKU + 副图
            operationId: BFF_GetProductDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/bff.v1.GetProductDetailResponse'
    /api/v1/bff/shops/{id}:
        get:
            tags:
                - BFF
            description: 店铺主页 — 聚合店铺 + 商品列表
            operationId: BFF_GetShopHome
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/bff.v1.GetShopHomeResponse'
...
    /api/v1/products:
        get:
            tags:
                - Product
            description: 分页查询商品列表（含店铺信息）
            operationId: Product_ListProducts
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: shopId
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.ListProductsResponse'
        post:
            tags:
                - Product
            description: 创建商品
            operationId: Product_CreateProduct
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/product.v1.CreateProductRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.CreateProductResponse'
    /api/v1/products/{id}:
        get:
            tags:
                - Product
            description: 查看商品详情（含店铺信息）
            operationId: Product_GetProduct
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.GetProductResponse'
        put:
            tags:
                - Product
            description: 更新商品
            operationId: Product_UpdateProduct
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/product.v1.UpdateProductRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.UpdateProductResponse'
        delete:
            tags:
                - Product
            description: 删除商品
            operationId: Product_DeleteProduct
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product.v1.DeleteProductResponse'
...
components:
    schemas:
        bff.v1.GetProductDetailResponse:
            type: object
            properties:
                product:
                    $ref: '#/components/schemas/bff.v1.ProductDetail'
        bff.v1.GetShopHomeResponse:
            type: object
            properties:
                shop:
                    $ref: '#/components/schemas/bff.v1.ShopHome'
        bff.v1.ListProductsResponse:
            type: object
            properties:
                products:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.ProductListItem'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        bff.v1.Media:
            type: object
            properties:
                id:
                    type: string
                url:
                    type: string
                sort:
                    type: integer
                    format: int32
        bff.v1.ProductDetail:
            type: object
            properties:
                id:
                    type: string
                shopId:
                    type: string
                shopName:
                    type: string
                name:
                    type: string
                type:
                    type: string
                description:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                sort:
                    type: integer
                    format: int32
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.Tag'
                medias:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.Media'
                skus:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.Sku'
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 商品详情页聚合数据
        bff.v1.ProductListItem:
            type: object
            properties:
                id:
                    type: string
                shopId:
                    type: string
                shopName:
                    type: string
                name:
                    type: string
                type:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.Tag'
                createdAt:
                    type: string
            description: 商品列表项（精简版）
        bff.v1.ShopHome:
            type: object
            properties:
                id:
                    type: string
                shopName:
                    type: string
                description:
                    type: string
                products:
                    type: array
                    items:
                        $ref: '#/components/schemas/bff.v1.ProductListItem'
            description: 店铺主页数据
        bff.v1.Sku:
            type: object
            properties:
                id:
                    type: string
                sku:
                    type: string
                title:
                    type: string
                price:
                    type: string
                stock:
                    type: integer
                    format: int32
                imgUrl:
                    type: string
        bff.v1.Tag:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
...
        product.v1.CreateProductRequest:
            type: object
            properties:
                shopId:
                    type: string
                name:
                    type: string
                type:
                    type: string
                description:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                sort:
                    type: integer
                    format: int32
                tagIds:
                    type: array
                    items:
                        type: integer
                        format: int32
            description: 创建商品
        product.v1.CreateProductResponse:
            type: object
            properties:
                product:
                    $ref: '#/components/schemas/product.v1.ProductInfo'
        product.v1.DeleteProductResponse:
            type: object
            properties:
                success:
                    type: boolean
        product.v1.GetProductResponse:
            type: object
            properties:
                product:
                    $ref: '#/components/schemas/product.v1.ProductInfo'
        product.v1.ListProductsResponse:
            type: object
            properties:
                products:
                    type: array
                    items:
                        $ref: '#/components/schemas/product.v1.ProductInfo'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        product.v1.ProductInfo:
            type: object
            properties:
                id:
                    type: string
                shopId:
                    type: string
                shopName:
                    type: string
                name:
                    type: string
                type:
                    type: string
                description:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                sort:
                    type: integer
                    format: int32
                tagIds:
                    type: array
                    items:
                        type: integer
                        format: int32
                tagNames:
                    type: array
                    items:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 商品实体（API 返回结构）
        product.v1.UpdateProductRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                type:
                    type: string
                description:
                    type: string
                mainImageUrl:
                    type: string
                price:
                    type: string
                compareAtPrice:
                    type: string
                status:
                    type: integer
                    format: int32
                sort:
                    type: integer
                    format: int32
                tagIds:
                    type: array
                    items:
                        type: integer
                        format: int32
            description: 更新商品
        product.v1.UpdateProductResponse:
            type: object
            properties:
                product:
                    $ref: '#/components/schemas/product.v1.ProductInfo'
...